Reorganized into more logical units
