+++ common/CMakeLists.txt
@@ -86,6 +86,7 @@ set(common_sources
     packet_dump_file.cpp
     packet_dump.cpp
     packet_functions.cpp
+	patches/client_version.cpp
     patches/patches.cpp
     patches/rof_limits.cpp
     patches/rof.cpp
@@ -655,6 +656,8 @@ set(common_headers
     packet_dump_file.h
     packet_dump.h
     packet_functions.h
+	patches/IMessage.h
+	patches/client_version.h
     patches/patches.h
     patches/rof_limits.h
     patches/rof_ops.h
