--- zone/patch/components/CMakeLists.txt
@@ -1,1 +1,0 @@
-add_subdirectory(message)
